+++ Controls/dataSource.ts
@@ -20,3 +20,4 @@ export {
     IControllerOptions as ISourceControllerOptions
 } from './_dataSource/Controller';
 export {default as calculatePath} from 'Controls/_dataSource/calculatePath';
+export {isEqualItems} from './_dataSource/Controller';
